separated plugins options into individual files
