+++ nvim/lua/plugins/comments.lua
@@ -1,0 +1,1 @@
+require("Comment").setup()
