+++ nvim/lua/plugins/floaterm.lua
@@ -1,0 +1,8 @@
+vim.g.floaterm_autoclose = 2
+vim.g.floaterm_keymap_toggle = "<F2>"
+vim.g.floaterm_width = 0.85
+vim.g.floaterm_height = 0.9
+vim.g.floaterm_rootmarkers = { ".git" }
+vim.g.floaterm_opener = "edit"
+vim.g.floaterm_borderchars = "        "
+vim.g.floaterm_title = ""
