+++ nvim/lua/plugins/inc_rename.lua
@@ -1,0 +1,1 @@
+require("inc_rename").setup()
