+++ nvim/lua/plugins/lastplace.lua
@@ -1,0 +1,5 @@
+require("nvim-lastplace").setup({
+	lastplace_ignore_buftype = { "quickfix", "nofile", "help" },
+	lastplace_ignore_filetype = { "gitcommit", "gitrebase", "svn", "hgcommit" },
+	lastplace_open_folds = true,
+})
