+++ nvim/lua/plugins/marks.lua
@@ -1,0 +1,10 @@
+require("marks").setup({
+	mappings = {
+		set_next = "mm",
+		next = "mn",
+		prev = "mp",
+		preview = false,
+	},
+})
+
+vim.keymap.set("n", "m/", ":MarksListAll<CR>")
