+++ nvim/lua/plugins/neoscroll.lua
@@ -1,0 +1,3 @@
+require("neoscroll").setup({
+	mappings = { "<C-u>", "<C-d>", "<C-e>", "zt", "zz", "zb" },
+})
