+++ nvim/lua/plugins/notify.lua
@@ -1,0 +1,5 @@
+require("notify").setup({
+	timeout = 5000,
+	render = "minimal",
+	stages = "slide",
+})
