+++ nvim/lua/plugins/quickfix.lua
@@ -1,0 +1,5 @@
+require("bqf").setup({ func_map = { openc = "<CR>" } })
+require("pqf").setup({
+	signs = { error = "✘", warning = ".", info = "i", hint = "💡" },
+})
+require("nvim-jqx.config").use_quickfix = false
