+++ nvim/lua/plugins/symbols_outline.lua
@@ -1,0 +1,7 @@
+require("symbols-outline").setup({
+	highlight_hovered_item = false,
+	autofold_depth = 1,
+	symbols = {
+		Function = { icon = Icons.Function },
+	},
+})
