+++ nvim/lua/plugins/treesitter.lua
@@ -1,0 +1,4 @@
+require("nvim-treesitter.configs").setup({
+	highlight = { enable = true },
+	ensure_installed = { "go", "python", "json", "lua", "bash", "regex", "yaml", "vim" },
+})
