+++ nvim/lua/plugins/vimtex.lua
@@ -1,0 +1,6 @@
+vim.g.vimtex_view_general_viewer = "zathura"
+vim.g.vimtex_view_zathura_options = "-reuse-instance"
+vim.g.vimtex_imaps_leader = ","
+vim.g.tex_conceal = ""
+vim.g.tex_fast = ""
+vim.g.tex_flavor = "latex"
